Fix the case where tasks use `ToolingModelBuilderRegistry` to locate and build a tooling model.
Introduce an internal service that is used by the tooling API entry points to locate a tooling model builder that can build a tooling model as a top level build action (that is, that fires the appropriate build operations, sets up the worker thread state, and so on). The public API method now returns a tooling model builder that simply builds the model and that can be used by build logic to build models.